fix TARGET_FILTER corner case for SDKs
If a TARGET_FILTER is set only imagebuilder/sdk/rootfs for that target
are created. In case of the SDK multiple valid targets are in the same
row printed via `dump-target-info.pl`. With the previous logic the
"first" target is used expecting all targets to be on the same commit.
However when using TARGET_FILTER the specific target should be used for
the SDK, not just one that supports the same architecture.

A logic is added that if TARGET_FILTER is set, it's used as the SDK
target.

Signed-off-by: Paul Spooren <mail@aparcar.org>